# Improve JsonObjectParser performance