In our application, users (other apps) are uploading big (few GBs) files with Json lines. Each document is around 50kB
To split them I use `JsonFraming.objectScanner`
When profiling that code I noticed many calls to `scala.runtime.BoxesRunTime`:
![screen shot 2018-06-21 at 16 29 47](https://user-images.githubusercontent.com/177043/41726098-b00c99f6-7571-11e8-8492-02049b3ca2ea.png)

This PR replaces usage of `Set.contains` with normal pattern match, and improves overall performance for both small and large documents.

```
  Benchmark                                     Mode  Cnt        Score       Error  Units
    //old
    JsonFramingBenchmark.counting_1              thrpt   25  4255350.000 ± 18824.499  ops/s
    JsonFramingBenchmark.counting_long_document  thrpt   25      344.765 ±     0.452  ops/s
    JsonFramingBenchmark.counting_offer_5        thrpt   25  4501980.800 ± 11976.386  ops/s

    // new
    JsonFramingBenchmark.counting_1              thrpt   25  5654469.050 ±  9360.659  ops/s
    JsonFramingBenchmark.counting_long_document  thrpt   25      425.550 ±     0.524  ops/s
    JsonFramingBenchmark.counting_offer_5        thrpt   25  5300042.946 ± 24006.358  ops/s
```


Also I noticed that `JsonFramingBenchmark` is broken on `master` as it leaks memory in every run - more data is `offered` than `polled` causing internal buffer to grow.