@lustefaniak 2018-06-21
What should I do with 
```
[error]  * method isWhitespace(Byte)Boolean in object akka.stream.impl.JsonObjectParser does not have a correspondent in current version
[error]    filter with: ProblemFilters.exclude[DirectMissingMethodProblem]("akka.stream.impl.JsonObjectParser.isWhitespace")
[error]  * method Whitespace()scala.collection.immutable.Set in object akka.stream.impl.JsonObjectParser does not have a correspondent in current version
[error]    filter with: ProblemFilters.exclude[DirectMissingMethodProblem]("akka.stream.impl.JsonObjectParser.Whitespace")
```
`JsonObjectParser` is marked `@InternalApi private[akka]` so I don't think this is a problem?