@johanandren 2018-06-22
The mima warnings seems safe to put on ignore list. See https://github.com/akka/akka/blob/master/CONTRIBUTING.md#binary-compatibility for details