@lustefaniak 2018-06-22
Added mima ignores