@lustefaniak 2018-06-22
Fail on travis looks like cache corruption: 
```
 java.io.IOException: Error accessing /home/travis/.ivy2/cache/com.typesafe.akka/akka-stream_2.12/jars/akka-stream_2.12-2.5.6.jar
```