@ktoso 2018-06-26
Good in general, some thoughts that the val is not the best there perhaps... Would have done a `def + match` I'll check if it matters